Adding categorizer feature
This feature allows user to add a Category column to a dataset using a dropdown on the Streamlit app and by using a predefined set of categories from a separate file which can be updated by the User to add or remove a list of categories. Also, allowing Users to download the file after updating the Categories.